Set several fields as optional
StoredFile::position_in_archive and
StoredFile::position_in_central_directory
have been set optional because it's possible to find a file
only in archive or only in central directory. For example, if
the field StoredFile::found_in_archive is set to false, there is
no reason to have a value for StoredFile::position_in_archive.
An absent file has no position.

StoredFile::central_directory has been set optional, it doesn't match
the specification. In case of a damaged ZIP file without central
directory, it could still be possible to read some stored files from
the archive. For this to happen it's required to be able to manage
partial ZIP files, without central directory.