Imprement missing structures
Several optional parts of a ZIP like Archive Decryption Header have
been represented as structs.

In the meantime, StoredFile evolved. A StoredFile represents a file as
stored in the first parts of a ZIP, so it's not the right place to store
the position and offset of the file as found in the central directory.
These values are stored in the CentralDirectoryFileHeader struct.
However, StoredFile has a boolean field to know whether the file exists
in the central directory, and CentralDirectoryFileHeader has a similar
field.