Implement missing readers
Most of the readers have been implemented, mainly about the central
directory. The encryption-related features of a ZIP file are not
handled yet, so the ArchiveDecryptionReader is not implemented.

Actually, the reading of a central directory doesn't work. The
end of central directory record signature is not found. Will
be fixed later. Currently the main prints the data it has, the
central directory is just ignored.