Improve the reading in case of damaged file
At first, local file headers are read one after the other.
Once it's not possible to read local file headers anymore, we don't
know if we reached the end of local file headers and are about to
start a new section (e.g. central directory), or if the file is
damaged. To avoid being stuck in this situation, the file cursor
is shift of 1 byte and the 4 next bytes are checked against the
known signatures. By doing this, as soon as we find another
known section, reading can continue.

However, central directory is still not detected.