Fix parsing of CentralDirectory
The central directory of a ZIP file is now detected. The central
directory headers and the end of central directory record are read
and their data are printed in main.

The parsing of central directory should be improved in the future:
actually it stops reading the file headers as soon as the signature
is no longer detected. That means that if a file header is damaged,
the next headers won't be read, neither the end of central directory
record.
The ZipFileReader::read() function is more flexible regarding damaged
file, it reads the file byte by byte until it finds a signature. In
the future, CentralDirectoryReader::read() will probably be more similar
to ZipFileReader::read().